Take out acceptInvite
